+++ src/textual/binding.py
@@ -53,6 +53,15 @@ class Binding:
     tooltip: str = ""
     """Optional tooltip to show in footer."""
 
+    def parse_key(self) -> tuple[list[str], str]:
+        """Parse a key in to a list of modifiers, and the actual key.
+
+        Returns:
+            A tuple of (MODIFIER LIST, KEY).
+        """
+        *modifiers, key = self.key.split("+")
+        return modifiers, key
+
 
 class ActiveBinding(NamedTuple):
     """Information about an active binding (returned from [active_bindings][textual.screen.Screen.active_bindings])."""
@@ -122,6 +131,14 @@ class BindingsMap:
         self.key_to_bindings: dict[str, list[Binding]] = {}
         for binding in make_bindings(bindings or {}):
             self.key_to_bindings.setdefault(binding.key, []).append(binding)
+
+    def _add_binding(self, binding: Binding) -> None:
+        """Add a new binding.
+
+        Args:
+            binding: New Binding to add.
+        """
+        self.key_to_bindings.setdefault(binding.key, []).append(binding)
 
     def __iter__(self) -> Iterator[tuple[str, Binding]]:
         """Iterating produces a sequence of (KEY, BINDING) tuples."""