+++ src/textual/command.py
@@ -436,7 +436,6 @@ class CommandPalette(SystemModalScreen):
     """
 
     DEFAULT_CSS = """
-    
    
     CommandPalette:inline {
         /* If the command palette is invoked in inline mode, we may need additional lines. */
@@ -575,6 +574,8 @@ class CommandPalette(SystemModalScreen):
         """Keeps track of if there are 'No matches found' message waiting to be displayed."""
         self._providers: list[Provider] = []
         """List of Provider instances involved in searches."""
+        self._hit_count: int = 0
+        """Number of hits displayed."""
 
     @staticmethod
     def is_open(app: App) -> bool:
@@ -625,7 +626,7 @@ class CommandPalette(SystemModalScreen):
         Returns:
             The content of the screen.
         """
-        with Vertical():
+        with Vertical(id="--container"):
             with Horizontal(id="--input"):
                 yield SearchIcon()
                 yield CommandInput(placeholder="Search for commands…")
@@ -883,6 +884,7 @@ class CommandPalette(SystemModalScreen):
         if highlighted is not None and highlighted.id:
             command_list.highlighted = command_list.get_option_index(highlighted.id)
         self._list_visible = bool(command_list.option_count)
+        self._hit_count = command_list.option_count
 
     _RESULT_BATCH_TIME: Final[float] = 0.25
     """How long to wait before adding commands to the command list."""
@@ -1014,6 +1016,7 @@ class CommandPalette(SystemModalScreen):
         # mean nothing was found. Give the user positive feedback to that
         # effect.
         if command_list.option_count == 0 and not worker.is_cancelled:
+            self._hit_count = 0
             self._start_no_matches_countdown(search_value)
 
     def _cancel_gather_commands(self) -> None:
@@ -1049,6 +1052,7 @@ class CommandPalette(SystemModalScreen):
         input.action_end()
         self._list_visible = False
         self.query_one(CommandList).clear_options()
+        self._hit_count = 0
         if self.run_on_select:
             self._select_or_command()
 
@@ -1063,10 +1067,15 @@ class CommandPalette(SystemModalScreen):
         if event is not None:
             event.stop()
         if self._list_visible:
+            command_list = self.query_one(CommandList)
             # ...so if nothing in the list is highlighted yet...
-            if self.query_one(CommandList).highlighted is None:
+            if command_list.highlighted is None:
                 # ...cause the first completion to be highlighted.
                 self._action_cursor_down()
+                # If there is one option, assume the user wants to select it
+                if command_list.option_count == 1:
+                    # Call after a short delay to provide a little visual feedback
+                    self._action_command_list("select")
             else:
                 # The list is visible, something is highlighted, the user
                 # made a selection "gesture"; let's go select it!
@@ -1090,12 +1099,9 @@ class CommandPalette(SystemModalScreen):
 
     def _action_escape(self) -> None:
         """Handle a request to escape out of the command palette."""
-        if self._list_visible:
-            self._list_visible = False
-        else:
-            self._cancel_gather_commands()
-            self.app.post_message(CommandPalette.Closed(option_selected=False))
-            self.dismiss()
+        self._cancel_gather_commands()
+        self.app.post_message(CommandPalette.Closed(option_selected=False))
+        self.dismiss()
 
     def _action_command_list(self, action: str) -> None:
         """Pass an action on to the [`CommandList`][textual.command.CommandList].